Add files via upload
The stuff in the src folder includes all the changes I've made. Add buttons and stuff as normal but don't remove the Router tags around the content in the app.js file or delete the routes in it. The Home component in the App.js file will be the main page. The About.js file is the About page I'm working on.